requirements.txt
dnspython
pyqt5